+++ NEWS.md
@@ -1,3 +1,5 @@
+# plu (development version)
+
 # plu 0.2.3
 
 * Added staticexport of `plu_ralize()`.