+++ NEWS.md
@@ -1,6 +1,8 @@
 # plu (development version)
 
 * Fixed an error where the plural forms of "octopus" and "platypus" were mixed up.
+* `plu_ral()` no longer attempts to replace the sequence `"n"`.
+  * The arguments `replace_n` and `n_fn` are now deprecated.
 * `plu_stick()` is now defunct.
 
 # plu 0.2.3